+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # Changelog
 
-## Next
+## 5.26.0 (1 March 2023)
 
 ### Features
 
 - `[cli]` - add `inject-extension` command that injects custom stats extension
 - `[cli]`: add `vrules` command that shows available validation rules that plugins in `statoscope.config.js` provides ([#192](https://github.com/statoscope/statoscope/pull/192) by @denisbalyko)
 
-## 5.25.1 (19 February 2022)
+## 5.25.1 (19 February 2023)
 
 ### Fixes
 
@@ -17,7 +17,7 @@
 - `[webpack-ui]` - module search now works only for first 500 chars in module name (performance issue)
 - `[webpack-ui]` - fix incorrect badge for asset modules
 
-## 5.25.0 (17 February 2022)
+## 5.25.0 (17 February 2023)
 
 ### Features